+++ nakadion/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.25.4] - 2020-05-14
+
+### CHANGED
+
+- histograms reset after configurable inactivity
+
+### ADDED
+
+- configurable histogram reset with `HistogramInactivityResetSecs`
+
 ## [0.25.3] - 2020-05-14
 
 ### CHANGED